backend/collector/requirements.txt
paramiko==4.0.0
mysql-connector-python==8.4.0